Create vboard-html.html
The HTML data for the Keyboard, this must be included in the body tag of your website, and the ids and classes must not be edited in order to run the keyboard. You can place it anywhere at the start of the body tag or at the end of it. Won't matter!